requirements.txt
torch
torchvision
ultralytics
opencv-python
Pillow
numpy